Edited Files: - CommunityChest.py - FreeParking.py - Tax.py
New Files:
- FreeParkingTotal.txt

Looked over the code to get an understanding.

Added some TODO's with comments on some possible issues/points I need clarification in CommunityChest.py.

Added FreeParkingTotal.txt to track the "Free Parking Pot" and functionality so landing on Free Parking adds the pot to the visitors balance and resets the pot.

Added functionality to Tax.py so that the tax money is added to the Free Parking pot.